Rename config_reader.rs to config.rs
