+++ site-builder/src/site.rs
@@ -5,6 +5,6 @@ pub mod builder;
 pub mod content;
 #[macro_use]
 pub mod macros;
-pub mod config_reader;
+pub mod config;
 pub mod manager;
 pub mod resource;
